requirements.txt
mcstatus
flask
pillow